third_party/WebKit/LayoutTests/fast/css3-text/css3-text-indent/getComputedStyle/getComputedStyle-text-indent-inherited-expected.txt

This test checks that the value of text-indent is properly inherited to the child.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Value of ancestor is '10px':
PASS window.getComputedStyle(ancestor).getPropertyValue('text-indent') is '10px'
PASS window.getComputedStyle(child).getPropertyValue('text-indent') is '10px'


Value of ancestor is '10px each-line':
PASS window.getComputedStyle(ancestor).getPropertyValue('text-indent') is '10px each-line'
PASS window.getComputedStyle(child).getPropertyValue('text-indent') is '10px each-line'


Value of ancestor is '10px hanging':
PASS window.getComputedStyle(ancestor).getPropertyValue('text-indent') is '10px hanging'
PASS window.getComputedStyle(child).getPropertyValue('text-indent') is '10px hanging'


Value of ancestor is '10px hanging each-line':
PASS window.getComputedStyle(ancestor).getPropertyValue('text-indent') is '10px hanging each-line'
PASS window.getComputedStyle(child).getPropertyValue('text-indent') is '10px hanging each-line'


Value of ancestor is '10px each-line', while child is '10px':
PASS window.getComputedStyle(ancestor).getPropertyValue('text-indent') is '10px each-line'
PASS window.getComputedStyle(child).getPropertyValue('text-indent') is '10px'


Value of ancestor is '10px hanging', while child is '10px':
PASS window.getComputedStyle(ancestor).getPropertyValue('text-indent') is '10px hanging'
PASS window.getComputedStyle(child).getPropertyValue('text-indent') is '10px'


Value of ancestor is '10px hanging each-line', while child is '10px':
PASS window.getComputedStyle(ancestor).getPropertyValue('text-indent') is '10px hanging each-line'
PASS window.getComputedStyle(child).getPropertyValue('text-indent') is '10px'


PASS successfullyParsed is true

TEST COMPLETE
